% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takeout_wrangling.R
\name{clean_locations}
\alias{clean_locations}
\title{Clean up Location History data}
\usage{
clean_locations(df)
}
\arguments{
\item{df}{Location history dataframe as output from json_to_tibble()}
}
\value{
Tibble
}
\description{
Rename columns, convert latitude and longitude notation.
}
