% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takeout_wrangling.R
\name{json_to_tibble}
\alias{json_to_tibble}
\title{Convert (potentially) nested .json to tibble}
\usage{
json_to_tibble(f)
}
\arguments{
\item{f}{Name of a .json file (string).}
}
\value{
Tibble
}
\description{
Read in a .json file, unnest and flatten it, and return as a tibble.
}
